// src/app/model/property.ts
export class property {
    userid: any;
    propertyCategoryId: any;
    propertyTypeId: any;
    propertyName: any;
    facing: any;
    price: any;
    description: any;
    propertyArea: any;
    constructionStatus: any;
    securityDeposit: any;
    maintainanceCost: any;
    rentalPeriod: any;
    communityId: any;
    propertyId: any;
    available: any;
    age: any;
    facilitypost: any;


    amenitypost: any;

    address: any;
    pincode: any;
    landmarks: any;
    countryId: any;
    stateId: any;
    cityId: any;
    areaId: any;

    latitude: any;
    longitude: any;

    propertyAmenityId: any;
    amenityValue: any;

    facilityId: any;

    image: any = [];

    videos: any = [];
    preference: any;
    ownername: any;
    ownermobile: any;
    ownermobile1: any;
    agentreferralid: any;
    updateddate: any;
    status: any;
    isfreead: any;
    bhktype: any;
    buildtype: any;
    vacantnoticeperiod: any;
    propertyvisittime: any;
    nooffloorsinbuilding: any;
    propertyinfloorno: any;
    floortype: any;
    bedrooms: any;
    restrooms: any;
    noiselevel: any;
    features: any;
    residencefurnishings: any;
    furnishedtype: any;
    commonfurnishings: any;
    sharetype: any;
    totalnoofrooms: any;
    totalnoofbeds: any;
    availablebeds: any;
    advanceamount: any;
    noticeperiod: any;
    foodtype: any;
    eventtype: any;
    halltype: any;
    booking_date: any;
    vegetarian: any;
    non_vegetarian: any;
    advance: any;
    averagepriceofroom: any;
    alcoholprovision: any;
    foodprovision: any;
    decorationprovision: any;
    otherpolicies: any;
    hosteltermsandpolicy: any;
    playgroundtermsandpolicy: any;
    hostelac: any;
    hostelacprice: any;
    hostelnonac: any;
    hostelnonacprice: any;
    eventfeaturesofvenue: any;
    hallprice: any;
    eventtypeprice: any;
    playgroundfeatures: any;
    vehiclefeatures: any;
    propertyshownby: any;
    propertyshownbynumber: any;
    smokingallowance: any;
    drinkingallowance: any;
    propertyrent: any;
    propertysecuritydeposit: any;

}


export class editing {
    image: any
    path: any
}

export class propertytype {
    propertyTypeId: any;
    radius: any;
    minprice: any;
    maxprice: any;
    lat: any;
    long: any;
    userid: any;
}